Removing unused/unwanted .desktop files
This includes both thunar and related files, as those cannot be removed
thanks to a dependency issue, as well as the COSMIC and Sway sessions.

The intended session is Niri, as that's what I use. I might do cleanup
work on the packages included once dnf5 has a module in Bluebuild.